+++ lib/flutter_web_scroll.dart
@@ -1,3 +1,4 @@
 export 'src/smooth_scroll_web.dart';
 export 'src/scroll_types.dart';
 export 'src/scroll_config.dart';
+